copy string in newly allocated memory

+++ 0x0B-malloc_free/1-strdup.c
@@ -6,7 +6,7 @@
  */
 char *_strdup(char *str)
 {
-	int m;
+	unsigned long int m;
 	char *p;
 	unsigned long int i = 0;
